# apps/api/openapi/v1.yaml
openapi: 3.0.0
paths:
...
  /api/v1/staff/{id}:
    patch:
      operationId: updateStaff
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Rename, change role, or (de)activate
      tags: *a2
  /api/v1/staff/{id}/pin:
    post:
      operationId: setStaffPin
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      summary: Set or reset a register PIN
      tags: *a2
...
info:
  title: RetailOS API
  description: REST API per 03-architecture/api-design.md. Internal clients are
    the first consumers; public from Phase 5.
  version: 1.0.0-phase0
  contact: {}
...
servers: []
...